+++ config.py
@@ -56,7 +56,7 @@ DIRECT_RECORDS = 'output/direct.json'
 LM_COL_CODE = 'Affcode'
 LM_COL_AFFL = 'Affil'
 LM_COLS = [LM_COL_CODE,LM_COL_AFFL]
-LM_TEXTLEN = 8192
+LM_TEXTLEN = 2048
 LM_THRESHOLD = -0.0
 
 MATCH_COL_AFFL = 'Affil'
