+++ tests/test_app.py
@@ -35,22 +35,6 @@ class TestLoadData(unittest.TestCase):
         self.assertTrue(filecmp.cmp(config.PICKLE_FILE,'tests/outputdata/test.pickle'))
 
 
-#this won't work as is, see Import_Pipeline's test_tasks for setup of celery worker
-#class TestDirectMatch(unittest.TestCase):
-#
-#    def test_augment_records(self):
-#        (aff_list,aff_dict,canon_list) = app.tasks.task_load_dicts_from_file(PC_INFILE,AFFDICT_INFILE)
-#        with open(DIRECT_RECORDS,'rU') as fj:
-#            jdata = json.load(fj)
-#            records = jdata['docs']
-#        self.assertEqual(len(records),4)
-#        ar = []
-#        ax = app.tasks.app_module.ADSAffilCelery('mock-celery','')
-#        for r in records:
-#            ar.append(ax.augment_affiliations(r))
-#        self.assertEqual(len(ar),4)
-#        self.assertEqual(ar[0],{'lol':'wut'})
-
 class TestDirectMatch(unittest.TestCase):
 
     def setUp(self):
@@ -74,7 +58,12 @@ class TestDirectMatch(unittest.TestCase):
 
     def test_direct_matching(self):
         with patch('ADSAffil.tasks.task_augment_affiliations_json', return_value=None) as next_task:
+#           global adict
+#           adict = app.utils.load_affil_dict(config.PICKLE_FILE)
             self.assertFalse(next_task.called) 
+#           in_rec1 = {"bibcode":u"2109zyxwv......12X", "aff":u"Harvard-Smithsonian Center for Astrophysics"}
+#           out_rec1 = self.app.augment_affiliations(in_rec1)
+#           self.assertEqual(in_rec1,out_rec1)
 
 
 class TestMachineLearning(unittest.TestCase):
@@ -83,3 +72,35 @@ class TestMachineLearning(unittest.TestCase):
         lmod = app.tasks.task_make_learning_model(aff_dict)
         app.tasks.task_resolve_unmatched(unmatched.keys(), lmod)
         self.assertTrue(filecmp.cmp(config.OUTPUT_FILE,'output/ml.out'))
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
